+++ packages/time/test/Polysemy/Time/Test/TimeUnitTest.hs
@@ -2,7 +2,7 @@ module Polysemy.Time.Test.TimeUnitTest where
 
 import Polysemy.Test (UnitTest, assertEq, runTestAuto)
 
-import Polysemy.Time.Data.TimeUnit (MicroSeconds (MicroSeconds), MilliSeconds (MilliSeconds))
+import Polysemy.Time.Data.TimeUnit (MicroSeconds (MicroSeconds), MilliSeconds (MilliSeconds), secondsFrac)
 
 test_fractional :: UnitTest
 test_fractional =
@@ -10,3 +10,5 @@ test_fractional =
     assertEq @_ @IO (MilliSeconds 50) 0.05
     assertEq @_ @IO (MicroSeconds 200) 0.0002
     assertEq @_ @IO (MilliSeconds 50) (MilliSeconds 100 / MilliSeconds 2)
+    assertEq @_ @IO (MilliSeconds 50) (MilliSeconds 100 / 2)
+    assertEq @_ @IO 0.05 (secondsFrac (MilliSeconds 50))