+++ packages/time/changelog.md
@@ -1,3 +1,6 @@
+# Unreleased
+* Add `secondsFrac`, a combinator that converts a unit into a fractional number of seconds.
+
 # 0.1.3.0
 * Added `measure`, a combinator that returns the time span an action takes.